macosx: fix C++11 build
backporting commit c3cf35ab63c04fb1d7b2f6760128f42c20cac0e1

+++ modules/highgui/src/window_cocoa.mm
@@ -266,7 +266,7 @@ CV_IMPL void cvResizeWindow( const char* name, int width, int height)
     CVWindow *window = cvGetWindow(name);
     if(window && ![window autosize]) {
         height += [window contentView].sliderHeight;
-        NSSize size = { width, height };
+        NSSize size = { (CGFloat)width, (CGFloat)height };
         [window setContentSize:size];
     }
     [localpool drain];
